requirements.txt
gradio>=4.26.0
kaggle>=1.6.0
huggingface-hub>=0.22.0
aiohttp>=3.9.0
requests>=2.31.0
faiss-cpu>=1.8.0
sentence-transformers>=2.7.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
tenacity>=8.2.0
huggingface_hub